+++ elements/kahoot-controller/src/app.tsx
@@ -1,9 +1,8 @@
-import { useState } from 'react'
 import './App.css'
-import { KahootPlayer } from './kahootPlayer'
-import { usePin } from './usePin'
-import { SettingScreen } from './settingScreen'
+import { usePin } from './hooks/usePin'
+import { KahootPlayer } from './screens/kahootPlayer'
 
+import { SettingScreen } from './screens/settingScreen'
 
 export function App() {
   const { pin, setPin } = usePin()
