--- elements/kahoot-controller/src/kahootPlayer.css
@@ -1,6 +1,0 @@
-.kahootPlayer {
-    width: 100%;
-    height: 100%;
-
-    border: none;
-}
